Add shutdown signal for loading threads
I don't know if this is silly or not. The threads are daemonic so they should be
ok to be killed roughly anyway. I added this for peace of mind that I don't have
lots of background threads doing things I don't want to do anymore, but I think
it may not be necessary.